Added user authorship check before habit update and delete

+++ backend/src/controllers/habitController.ts
@@ -26,6 +26,8 @@ export const create = async (req: Request<object, object, CreateHabitBody>, res:
 export const updateOne = async (req: Request<{ id: string }, object, UpdateHabitBody>, res: Response) => {
   const id = req.params.id;
   const data = req.body;
+  const userHabit = await habitService.findById({ _id: id, createdBy: res.locals.sub });
+  if (!userHabit) return res.status(304).json({ error: 'Incorrect user credentials' });
   try {
     const updatedHabit = await habitService.updateById(id, data);
     return res.status(200).json(updatedHabit);
@@ -37,6 +39,8 @@ export const updateOne = async (req: Request<{ id: string }, object, UpdateHabit
 
 export const deleteOne = async (req: Request<{ id: string }>, res: Response) => {
   const id = req.params.id;
+  const userHabit = await habitService.findById({ _id: id, createdBy: res.locals.sub });
+  if (!userHabit) return res.status(304).json({ error: 'Incorrect user credentials' });
   try {
     await habitService.deleteById(id);
     return res.status(204);
